pyproject.toml
[project]
name = "wheel-tracker"
version = "0.1.0"
description = "Wheel strategy tracker — FastAPI + SQLModel + Tailwind"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9",
    "jinja2>=3.1",
    "alembic>=1.14",
    "yfinance>=1.2.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
]
